+++ NEWS.md
@@ -1,5 +1,7 @@
 # ggiplot 0.0.1
 
-* Support multiple confidence levels (#2)
+* Tweaks to plot output, including integer breaks on x-axis (where appropriate)
+and allow additional user-level control (e.g. CI alpha or width levels)
+* Support multiple confidence levels (#2, #5)
 * Support multiple LHS variables (#1)
 * Added a `NEWS.md` file to track changes to the package.